+++ global_trotter_web/jest.config.cjs
@@ -5,5 +5,7 @@ module.exports = {
   moduleNameMapper: {
     '\\.(css|less|scss)$': 'identity-obj-proxy',
     '\\.(jpg|jpeg|png|gif|svg|webp)$': '<rootDir>/tests/__mocks__/fileMock.cjs'
-  }
+  },
+  maxWorkers: 2,
+  workerIdleMemoryLimit: '512MB'
 }